fix: source modal closing after popup
