+++ packages/shared/src/components/modals/NewSourceModal.tsx
@@ -2,7 +2,6 @@ import React, { ReactElement, useContext, useRef, useState } from 'react';
 import { useMutation } from 'react-query';
 import classNames from 'classnames';
 import request from 'graphql-request';
-import { cloudinary } from '../../lib/image';
 import { Button } from '../buttons/Button';
 import { ModalProps as StyledModalProps } from './StyledModal';
 import { SearchField } from '../fields/SearchField';
@@ -20,9 +19,7 @@ import AuthContext from '../../contexts/AuthContext';
 import { AuthTriggers } from '../../lib/auth';
 import { Modal, ModalProps } from './common/Modal';
 import NotificationsContext from '../../contexts/NotificationsContext';
-import { Justify } from '../utilities';
-import { useAnalyticsContext } from '../../contexts/AnalyticsContext';
-import { AnalyticsEvent, Origin } from '../../lib/analytics';
+import PushNotificationModal from './PushNotificationModal';
 
 interface RSS {
   url: string;
@@ -54,12 +51,11 @@ type ScrapeSourceResponse =
 
 export default function NewSourceModal(props: StyledModalProps): ReactElement {
   const scrapeFormRef = useRef<HTMLFormElement>();
-  const { trackEvent } = useAnalyticsContext();
   const [enableSubmission, setEnableSubmission] = useState(false);
   const [scrapeError, setScrapeError] = useState<string>();
   const [showContact, setShowContact] = useState(false);
   const [showNotification, setShowNotification] = useState(false);
-  const { isSubscribed, isNotificationSupported, onTogglePermission } =
+  const { hasPermissionCache, isNotificationSupported } =
     useContext(NotificationsContext);
   const [feeds, setFeeds] = useState<{ label: string; value: string }[]>();
   const [selectedFeed, setSelectedFeed] = useState<string>();
@@ -67,17 +63,6 @@ export default function NewSourceModal(props: StyledModalProps): ReactElement {
   const { user, loginState, showLogin } = useContext(AuthContext);
   const loginTrigger = AuthTriggers.SubmitNewSource;
   const { onRequestClose } = props;
-
-  const enableNotifications = async () => {
-    const permission = await onTogglePermission();
-    trackEvent({
-      event_name: AnalyticsEvent.ClickEnableNotification,
-      extra: JSON.stringify({ origin: Origin.NewSourceModal, permission }),
-    });
-    if (permission === 'granted') {
-      onRequestClose?.(null);
-    }
-  };
 
   const failedToScrape = () => {
     setShowContact(true);
@@ -139,7 +124,7 @@ export default function NewSourceModal(props: StyledModalProps): ReactElement {
         }),
       {
         onSuccess: () => {
-          if (isSubscribed || !isNotificationSupported) {
+          if (hasPermissionCache || !isNotificationSupported) {
             onRequestClose?.(null);
             return;
           }
@@ -179,34 +164,12 @@ export default function NewSourceModal(props: StyledModalProps): ReactElement {
     }
   };
   const modalProps: Omit<ModalProps, 'children'> = {
-    kind: Modal.Kind.FlexibleTop,
-    size: Modal.Size.Medium,
     onRequestClose,
     ...props,
   };
 
-  if (showNotification) {
-    return (
-      <Modal {...modalProps}>
-        <Modal.Header />
-        <Modal.Body>
-          <Modal.Title>Push notifications</Modal.Title>
-          <Modal.Text className="text-center">
-            Get notified on the status of your source submissions in real time
-          </Modal.Text>
-          <img
-            className="my-14 mx-auto"
-            src={cloudinary.notifications.big}
-            alt="A sample browser notification"
-          />
-        </Modal.Body>
-        <Modal.Footer justify={Justify.Center}>
-          <Button className="btn-primary" onClick={enableNotifications}>
-            Enable notifications
-          </Button>
-        </Modal.Footer>
-      </Modal>
-    );
+  if (!showNotification) {
+    return <PushNotificationModal {...modalProps} />;
   }
 
   return (
