+++ packages/shared/src/components/modals/common/Modal.tsx
@@ -14,13 +14,13 @@ import {
 } from './types';
 import classed from '../../../lib/classed';
 
-export type ModalProps = ReactModal.Props & {
+export interface ModalProps extends ReactModal.Props {
   children: React.ReactNode;
   kind?: ModalKind;
   size?: ModalSize;
   tabs?: string[] | ModalTabItem[];
   defaultTab?: string;
-};
+}
 
 const modalKindToOverlayClassName: Record<ModalKind, string> = {
   [ModalKind.FixedCenter]: 'mobileL:justify-center pt-10 mobileL:pt-0',
